Move o suporte do reactor para um novo projeto
