+++ java-restify-reactor/src/main/java/com/github/ljtfreitas/restify/http/client/call/exec/reactor/FluxEndpointCallExecutableFactory.java
@@ -23,7 +23,7 @@
  * SOFTWARE.
  *
  *******************************************************************************/
-package com.github.ljtfreitas.restify.http.spring.client.call.exec;
+package com.github.ljtfreitas.restify.http.client.call.exec.reactor;
 
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
