+++ java-restify-reactor/src/main/resources/META-INF/services/com.github.ljtfreitas.restify.http.client.call.exec.EndpointCallExecutableProvider
@@ -1,0 +1,2 @@
+com.github.ljtfreitas.restify.http.client.call.exec.reactor.MonoEndpointCallExecutableFactory
+com.github.ljtfreitas.restify.http.client.call.exec.reactor.FluxEndpointCallExecutableFactory
