+++ java-restify-reactor/src/test/java/com/github/ljtfreitas/restify/http/client/call/exec/reactor/MonoEndpointCallExecutableFactoryTest.java
@@ -1,4 +1,4 @@
-package com.github.ljtfreitas.restify.http.spring.client.call.exec;
+package com.github.ljtfreitas.restify.http.client.call.exec.reactor;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
