--- java-restify-spring-reactive/src/main/resources/META-INF/services/com.github.ljtfreitas.restify.http.client.call.exec.EndpointCallExecutableProvider
@@ -1,2 +1,0 @@
-com.github.ljtfreitas.restify.http.spring.client.call.exec.MonoEndpointCallExecutableFactory
-com.github.ljtfreitas.restify.http.spring.client.call.exec.FluxEndpointCallExecutableFactory
